Split config.rs three ways, lift the picker out of app.rs, and say no to two trait changes
`config.rs` was 916 lines holding eleven unrelated settings structs, the layout
grid, path resolution, validation and migration hints. It is now three files
behind one flat namespace — the split is for readability, not a claim that
`[layout]` and `[weather]` are different concepts, so `crate::config::X` still
names everything:

  config/mod.rs      488  paths, loading, validation, state application
  config/widgets.rs  310  one settings struct per panel
  config/layout.rs   166  the grid, and the only part mirador writes back

This is where the missing `deny_unknown_fields` on the two theme structs hid
long enough to make its own migration hint unreachable, which is an argument
about file size rather than about serde.

The panel picker moves to `picker.rs`, taking 118 lines out of `app.rs`. It owns
its cursor and its drawing and nothing else: `handle_key` returns an `Action`
saying what was asked for, and the shell decides whether it can be done —
because a toggle means rewriting the layout, rebuilding every panel and putting
it all back when that fails. Handing the picker a `&mut App` would have moved
the code without moving the responsibility. The cursor tests move with it, and
`app.rs` keeps the integration ones, which is what they were really testing.

Two recommendations from the review are **declined**, with the reasoning written
where the next person will look for it rather than in a commit nobody reads.

`describe() -> PanelInfo` would take the trait from thirteen methods to eight by
bundling the six "describe yourself" ones. It would also be slower and less
useful. The six are asked for at very different rates — `max_width` during
layout, `refresh_interval` on the tick, `bindings` in three places, and `title`
and `counter` on *every frame* from the render loop, with nothing guarding
them. One `describe()` computes all six whenever any one is wanted and returns
an owned struct holding a `String` and an `Option<String>`, so the cheapest
question would allocate twice per panel per frame. The trait is wide because the
shell's questions are separate. Noted in `panel.rs`.

Folding `captures_input` into a `KeyOutcome::ConsumedModal` cannot work: two of
its three call sites have no key outcome to read. The resize keys are claimed by
the shell *before* the panel is offered anything, so a panel mid-form has to veto
`Ctrl+Left` without having been called; and a mouse click must not pull focus out
of a half-typed task, with no key event anywhere in that path. The third looks
foldable and is not — the veto applies precisely when the panel returned
`Ignored`, because a form ignores `q` as not-one-of-its-keys and `q` must still
not quit. Also noted in `panel.rs`.

Verified under tmux: opened the picker, moved the cursor, toggled `todo` off and
watched both the mark and the panel go. 429 tests pass; fmt, clippy, doc and
MSRV green.
