+++ src/config/mod.rs
@@ -1,5 +1,11 @@
 //! Configuration loading.
 //!
+//! Split three ways: this file resolves the path, reads the file and
+//! validates the result; [`layout`] holds the grid; [`widgets`] holds one
+//! settings struct per panel. They were one 916-line module, which is where a
+//! misspelled `[theme]` key hid long enough to make its own migration hint
+//! unreachable.
+//!
 //! Mirador reads a single TOML file. On first run, if no file exists, a fully
 //! commented default is written to disk so there is always something to edit.
 //!
@@ -16,7 +22,26 @@ use serde::Deserialize;
 use crate::theme::Theme;
 
 /// The default config written on first run.
-pub const DEFAULT_CONFIG: &str = include_str!("../assets/default_config.toml");
+pub const DEFAULT_CONFIG: &str = include_str!("../../assets/default_config.toml");
+
+mod layout;
+mod widgets;
+
+// One flat namespace: the split into files is for readability, not a claim
+// that `[layout]` and `[weather]` are different concepts. Everything a reader
+// or a widget names lives at `crate::config::`.
+//
+// `LayoutPanel`, `LayoutRow` and `ClockZone` are named only by tests today —
+// production code builds them through serde and reaches them through their
+// parents — so the non-test build sees the re-export as unused. That is a fact
+// about who *names* the type, not about whether it is part of the surface.
+#[allow(unused_imports)]
+pub use layout::{Layout, LayoutPanel, LayoutRow};
+#[allow(unused_imports)]
+pub use widgets::{
+    CalendarConfig, ClockZone, ClocksConfig, CpuConfig, NetworkConfig, NotesConfig, PomodoroConfig,
+    StocksConfig, TodoConfig, WeatherConfig,
+};
 
 /// Top-level configuration.
 #[derive(Debug, Clone, Deserialize, Default)]
@@ -62,459 +87,6 @@ impl Default for General {
             show_borders: true,
             show_status_bar: true,
             mouse: true,
-        }
-    }
-}
-
-/// A grid of rows, each holding one or more side-by-side panels.
-///
-/// `height` and `width` are relative weights, not absolute cells, so a layout
-/// keeps its proportions at any terminal size.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct Layout {
-    pub rows: Vec<LayoutRow>,
-}
-
-impl Default for Layout {
-    fn default() -> Self {
-        // Kept identical to the `[layout]` block in assets/default_config.toml,
-        // and `the_rust_default_layout_matches_the_shipped_one` fails if they
-        // drift. They are reached by different routes — the shipped file on a
-        // true first run, this on any config that omits `[layout]` — and when
-        // they disagreed, deleting the section silently cost three panels.
-        let row = |height: u16, panels: &[(&str, u16)]| LayoutRow {
-            height,
-            panels: panels
-                .iter()
-                .map(|(widget, width)| LayoutPanel {
-                    widget: (*widget).into(),
-                    width: *width,
-                })
-                .collect(),
-        };
-
-        Self {
-            rows: vec![
-                row(34, &[("clocks", 26), ("calendar", 34), ("weather", 40)]),
-                row(42, &[("todo", 48), ("notes", 30), ("pomodoro", 22)]),
-                row(24, &[("stocks", 40), ("cpu", 30), ("network", 30)]),
-            ],
-        }
-    }
-}
-
-impl Layout {
-    /// Every widget this layout places, in row-major order.
-    pub fn widgets(&self) -> Vec<&str> {
-        self.rows
-            .iter()
-            .flat_map(|row| row.panels.iter())
-            .map(|panel| panel.widget.as_str())
-            .collect()
-    }
-
-    /// Whether `widget` appears anywhere.
-    pub fn places(&self, widget: &str) -> bool {
-        self.widgets().contains(&widget)
-    }
-
-    /// Place `widget`, appending it to the row carrying the fewest panels.
-    ///
-    /// A dashboard has no obvious "end" to add to, so the rule is the one that
-    /// keeps the grid even: the emptiest row, and the last of those on a tie so
-    /// repeated additions walk downward rather than piling into row one. The
-    /// new panel takes the average weight of that row's existing panels, which
-    /// leaves the others in proportion to each other instead of squeezing one.
-    ///
-    /// Deliberately crude, because it has to be undoable by hand: this only
-    /// ever *adds*, and `Ctrl+arrow` or the config can put it where you want.
-    /// Guessing more cleverly would be harder to predict, not easier.
-    pub fn add_widget(&mut self, widget: &str) {
-        if self.places(widget) {
-            return;
-        }
-
-        let target = self
-            .rows
-            .iter()
-            .enumerate()
-            .min_by_key(|(index, row)| (row.panels.len(), usize::MAX - index))
-            .map(|(index, _)| index);
-
-        let Some(index) = target else {
-            // No rows at all: give the widget one of its own rather than
-            // dropping it silently.
-            self.rows.push(LayoutRow {
-                height: 100,
-                panels: vec![LayoutPanel {
-                    widget: widget.to_string(),
-                    width: 100,
-                }],
-            });
-            return;
-        };
-
-        let row = &mut self.rows[index];
-        let width = if row.panels.is_empty() {
-            100
-        } else {
-            let total: u32 = row.panels.iter().map(|p| u32::from(p.width)).sum();
-            u16::try_from(total / row.panels.len() as u32)
-                .unwrap_or(25)
-                .max(1)
-        };
-        row.panels.push(LayoutPanel {
-            widget: widget.to_string(),
-            width,
-        });
-    }
-
-    /// Remove every placement of `widget`, and any row left empty by it.
-    ///
-    /// Refuses to remove the last panel on the dashboard and says so, because
-    /// an empty layout is rejected at startup — turning off the final panel
-    /// would produce a config that cannot be loaded next time.
-    pub fn remove_widget(&mut self, widget: &str) -> bool {
-        if self.widgets().iter().all(|placed| *placed == widget) {
-            return false;
-        }
-        for row in &mut self.rows {
-            row.panels.retain(|panel| panel.widget != widget);
-        }
-        self.rows.retain(|row| !row.panels.is_empty());
-        true
-    }
-}
-
-/// One horizontal band of the dashboard.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct LayoutRow {
-    /// Relative height weight.
-    pub height: u16,
-    pub panels: Vec<LayoutPanel>,
-}
-
-impl Default for LayoutRow {
-    fn default() -> Self {
-        Self {
-            height: 1,
-            panels: Vec::new(),
-        }
-    }
-}
-
-/// One panel within a row.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct LayoutPanel {
-    /// Widget id: `clocks`, `weather`, `todo`, `cpu` or `network`.
-    pub widget: String,
-    /// Relative width weight.
-    pub width: u16,
-}
-
-impl Default for LayoutPanel {
-    fn default() -> Self {
-        Self {
-            widget: String::new(),
-            width: 1,
-        }
-    }
-}
-
-/// World clock settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct ClocksConfig {
-    /// Clocks to display, in order.
-    pub zones: Vec<ClockZone>,
-    /// `strftime`-style time format.
-    pub time_format: String,
-    /// `strftime`-style date format. Empty hides the date.
-    pub date_format: String,
-    /// Show each zone's offset relative to the primary clock.
-    pub show_offset: bool,
-    /// Include seconds in the large clock. Off by default: a ticking seconds
-    /// field draws the eye every second, which is the opposite of what a
-    /// leave-it-running dashboard wants.
-    pub show_seconds: bool,
-}
-
-impl Default for ClocksConfig {
-    fn default() -> Self {
-        Self {
-            zones: vec![
-                ClockZone {
-                    label: "Local".into(),
-                    timezone: "local".into(),
-                },
-                ClockZone {
-                    label: "UTC".into(),
-                    timezone: "UTC".into(),
-                },
-                ClockZone {
-                    label: "London".into(),
-                    timezone: "Europe/London".into(),
-                },
-                ClockZone {
-                    label: "Tokyo".into(),
-                    timezone: "Asia/Tokyo".into(),
-                },
-            ],
-            time_format: "%H:%M:%S".into(),
-            date_format: "%A %d %B".into(),
-            show_offset: true,
-            show_seconds: true,
-        }
-    }
-}
-
-/// A single clock.
-#[derive(Debug, Clone, Deserialize, Default)]
-#[serde(default, deny_unknown_fields)]
-pub struct ClockZone {
-    /// Display name.
-    pub label: String,
-    /// IANA timezone id, or the literal `local` for the system zone.
-    pub timezone: String,
-}
-
-/// Weather settings. Data comes from Open-Meteo, which needs no API key.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct WeatherConfig {
-    /// Place name, geocoded on startup when `latitude`/`longitude` are unset.
-    pub location: String,
-    /// Explicit latitude; skips geocoding.
-    pub latitude: Option<f64>,
-    /// Explicit longitude; skips geocoding.
-    pub longitude: Option<f64>,
-    /// `metric` (C, km/h) or `imperial` (F, mph).
-    pub units: String,
-    /// Number of forecast hours to show, 1 to 24.
-    pub forecast_hours: u8,
-    /// Minutes between refreshes.
-    pub refresh_minutes: u64,
-}
-
-impl Default for WeatherConfig {
-    fn default() -> Self {
-        Self {
-            location: "Boston, Massachusetts".into(),
-            latitude: None,
-            longitude: None,
-            units: "imperial".into(),
-            forecast_hours: 8,
-            refresh_minutes: 30,
-        }
-    }
-}
-
-/// To-do list settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct TodoConfig {
-    /// Path to the task file. `~` is expanded. Defaults to the data directory.
-    pub file: Option<PathBuf>,
-    /// Show completed tasks in the list.
-    pub show_completed: bool,
-    /// Initial sort: `smart`, `due`, `priority`, `created` or `title`.
-    pub sort: String,
-    /// Date format used in the list.
-    pub date_format: String,
-    /// Hide tasks whose due date is more than this many days out. 0 disables.
-    pub horizon_days: u32,
-}
-
-impl Default for TodoConfig {
-    fn default() -> Self {
-        Self {
-            file: None,
-            show_completed: false,
-            sort: "smart".into(),
-            date_format: "%a %d %b".into(),
-            horizon_days: 0,
-        }
-    }
-}
-
-/// Stock watchlist settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct StocksConfig {
-    /// Symbols used to seed the watchlist on first run only. After that the
-    /// watchlist file is the truth, so symbols added or removed in the UI stick.
-    pub symbols: Vec<String>,
-    /// Path to the watchlist file. `~` is expanded. Defaults to the data
-    /// directory. Only symbols are stored; prices are never written to disk.
-    pub file: Option<PathBuf>,
-    /// Where quotes come from. See `[stocks].source` in the default config.
-    pub source: String,
-    /// Seconds between polls. Clamped to a minimum of 60: the sources are free
-    /// and unauthenticated, and hammering them gets the address blocked.
-    pub refresh_secs: u64,
-    /// Milliseconds between individual symbol requests, so a watchlist goes out
-    /// as a trickle rather than a burst.
-    pub stagger_ms: u64,
-    /// Draw the intraday sparkline when the panel is wide enough for it.
-    pub show_sparkline: bool,
-}
-
-impl Default for StocksConfig {
-    fn default() -> Self {
-        Self {
-            symbols: vec!["AAPL".into(), "MSFT".into(), "^GSPC".into()],
-            file: None,
-            source: "yahoo".to_string(),
-            refresh_secs: 120,
-            stagger_ms: 400,
-            show_sparkline: true,
-        }
-    }
-}
-
-/// Notes settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct NotesConfig {
-    /// Path to the notes file. `~` is expanded. Defaults to the data directory.
-    pub file: Option<PathBuf>,
-    /// Date format used in the list.
-    pub date_format: String,
-    /// Where the note body sits: `below` the list, or `beside` it.
-    ///
-    /// Below by default. Beside splits a finite width between two things that
-    /// both want it — the list loses room for titles and the body loses room
-    /// for prose — where stacking gives each the full width and trades only
-    /// height, which is the cheaper axis for both.
-    pub preview: String,
-}
-
-impl Default for NotesConfig {
-    fn default() -> Self {
-        Self {
-            file: None,
-            date_format: "%d %b".to_string(),
-            preview: "below".to_string(),
-        }
-    }
-}
-
-/// Calendar settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct CalendarConfig {
-    /// How many months to show, starting with the current one. The panel draws
-    /// as many as its size allows, up to this number.
-    pub months: u8,
-    /// `sunday` or `monday`.
-    pub week_starts: String,
-}
-
-impl Default for CalendarConfig {
-    fn default() -> Self {
-        Self {
-            months: 2,
-            week_starts: "sunday".to_string(),
-        }
-    }
-}
-
-/// Pomodoro timer settings.
-///
-/// These are the *starting* values. `+` and `-` change the timer in the panel,
-/// and those changes are remembered in the state file beside your tasks.
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-#[serde(default, deny_unknown_fields)]
-pub struct PomodoroConfig {
-    /// Length of a focus interval, in minutes.
-    pub focus_minutes: u64,
-    /// Length of the break after an ordinary focus interval.
-    pub short_break_minutes: u64,
-    /// Length of the break that closes a set.
-    pub long_break_minutes: u64,
-    /// Focus intervals per set, after which the long break falls due.
-    pub rounds_before_long_break: u32,
-    /// Begin the next phase the moment the current one ends, rather than
-    /// waiting for a keypress.
-    pub auto_start: bool,
-    /// Sound a notification when a phase ends. Off by default: a dashboard you
-    /// leave open all day has no business making noise you did not ask for.
-    pub chime: bool,
-    /// What to run for that notification, as a program and its arguments.
-    ///
-    /// Empty means the terminal bell, which costs nothing and lets your
-    /// terminal and OS decide whether that is a sound, a flash, or nothing at
-    /// all. Set it to play an actual file if you want a specific chime — see
-    /// the commented examples in the default config.
-    ///
-    /// Run directly rather than through a shell, so there is no quoting to get
-    /// wrong and no shell to inject into.
-    pub chime_command: Vec<String>,
-}
-
-impl Default for PomodoroConfig {
-    fn default() -> Self {
-        Self {
-            focus_minutes: 25,
-            short_break_minutes: 5,
-            long_break_minutes: 15,
-            rounds_before_long_break: 4,
-            auto_start: false,
-            chime: false,
-            chime_command: Vec::new(),
-        }
-    }
-}
-
-/// CPU chart settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct CpuConfig {
-    /// Number of samples retained in the moving chart.
-    pub history: usize,
-    /// Seconds between samples.
-    pub sample_secs: u64,
-    /// Also draw a per-core breakdown when the panel is tall enough.
-    pub show_per_core: bool,
-    /// Percentage above which the readout turns the warning colour.
-    pub warn_pct: f32,
-    /// Percentage above which the readout turns the error colour.
-    pub critical_pct: f32,
-}
-
-impl Default for CpuConfig {
-    fn default() -> Self {
-        Self {
-            history: 120,
-            sample_secs: 1,
-            show_per_core: true,
-            warn_pct: 70.0,
-            critical_pct: 90.0,
-        }
-    }
-}
-
-/// Network chart settings.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct NetworkConfig {
-    /// Interfaces to include. Empty means every non-loopback interface.
-    pub interfaces: Vec<String>,
-    /// Number of samples retained in the moving chart.
-    pub history: usize,
-    /// Seconds between samples.
-    pub sample_secs: u64,
-}
-
-impl Default for NetworkConfig {
-    fn default() -> Self {
-        Self {
-            interfaces: Vec::new(),
-            history: 120,
-            sample_secs: 1,
         }
     }
 }
